Move new type classes to their own typing module
